+++ src/docs/pages/component/checkbox/index.vue
@@ -70,7 +70,7 @@
     [[[形态]]]
 
     :::preset
-    @name:formStatusWithStyle
+    @name:formStatusWithStyleAndSize
     @uikey:checkbox
     @defaultValue:['Tim']
     @attrs: :list="{Tim:'Tim Boelaars',Andrew:'Andrew Colin Beck',Victor:'Victor Erixon'}"